Clean: harmonizing & header-commenting code
