+++ src/drivers/framebuffer.c
@@ -1,11 +1,16 @@
+// Framebuffer driver
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include <stdint.h>
 #include "framebuffer.h"
 #include "serial.h"
 #include "../kernel/system.h"
 
-//extern uint32_t *g_framebuffer;
+extern char* framebuffer;
 
-void putpixel(uint32_t* fb, int pitch, int bpp, int x, int y, uint32_t color) // framebuffer pointer, x, y, color
+void putpixel(uint32_t* fb, int pitch, int bpp, int x, int y, uint32_t color)
 {
 	if (bpp == 32) {
 		uint32_t* pixel_addr = (uint32_t*)((uint8_t*)fb + y * pitch + x *(bpp / 8));
@@ -13,14 +18,6 @@ void putpixel(uint32_t* fb, int pitch, int bpp, int x, int y, uint32_t color) //
 	}
 }
 
-extern char* framebuffer;
-extern int scanline;
-extern char _binary_include_fonts_UniCyr_8x16_psf_start;
-uint16_t* unicode;
-
-#define PIXEL uint32_t
-
-// Character, cursor X, cursor Y, foreground, background
 void draw_char(unsigned short int c, int cx, int cy, uint32_t fg, uint32_t bg)
 {
 	PSF_font *font = (PSF_font*)&_binary_include_fonts_UniCyr_8x16_psf_start;
