+++ src/drivers/kb.c
@@ -1,13 +1,12 @@
+// Keyboard driver
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "../kernel/io.h"
 #include "../libc/stdio.h"
 #include "../kernel/system.h"
-
-#define KEYBOARD_BUFFER_SIZE 256
-
-#define LEFT_SHIFT_PRESSED 0x2A
-#define RIGHT_SHIFT_PRESSED 0x36
-#define LEFT_SHIFT_RELEASED 0xAA
-#define RIGHT_SHIFT_RELEASED 0xB6
+#include "kb.h"
 
 unsigned char kbdus[128] =
 {
