+++ src/drivers/serial.h
@@ -1,5 +1,10 @@
-#ifndef INCLUDE_SERIAL_H
-#define INCLUDE_SERIAL_H
+// Serial I/O driver header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
+#ifndef SERIAL_H
+#define SERIAL_H
 
 #define PORT 0x3f8 //COM1
 
@@ -9,4 +14,5 @@ void write_serial(const char a);
 void serial_puts(const char* str);
 void log(const char* str, const int errlevel);
 void serial_printf(int errlevel, const char* fmt, ...);
+
 #endif
