+++ src/drivers/timer.c
@@ -1,3 +1,8 @@
+// Programmable Interval Timer channel 0 driver
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "../kernel/system.h"
 #include "../libc/stdio.h"
