+++ src/kernel/gdt.c
@@ -1,21 +1,10 @@
+// Global descriptor table setup
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "gdt.h"
 #include "../libc/stdio.h"
-
-struct gdt_entry
-{
-	unsigned short limit_low;
-	unsigned short base_low;
-	unsigned char base_middle;
-	unsigned char access;
-	unsigned char granularity;
-	unsigned char base_high;
-} __attribute__((packed));
-
-struct gdt_ptr
-{
-	unsigned short limit;
-	unsigned int base;
-} __attribute__((packed));
 
 struct gdt_entry gdt[3];
 struct gdt_ptr gp;
