+++ src/kernel/irq.c
@@ -1,3 +1,8 @@
+// Interrupt Requests setup
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "system.h"
 #include "io.h"
 #include "idt.h"
