+++ src/kernel/isr.c
@@ -1,3 +1,8 @@
+// Interrupt service routines setup 
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "system.h"
 #include "../libc/stdio.h"
 #include "idt.h"
@@ -37,7 +42,6 @@ extern void isr31();
 
 void isr_install()
 {
-
 	idt_set_gate(0, (unsigned)isr0, 0x08, 0x8E);
 	idt_set_gate(1, (unsigned)isr1, 0x08, 0x8E);
 	idt_set_gate(2, (unsigned)isr2, 0x08, 0x8E);
@@ -116,4 +120,3 @@ void fault_handler(struct regs *r)
 		for (;;);
 	}
 }
-
