+++ src/kernel/kheap.h
@@ -1,3 +1,8 @@
+// Kernel heap management header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef KHEAP_H
 #define KHEAP_H
