+++ src/kernel/paging.h
@@ -1,12 +1,21 @@
+// Paging kernel module header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef PAGING_H
 #define PAGING_H
 
 #include "system.h"
 #include <stdint.h>
+
+#define INDEX_FROM_BIT(a) (a/(8*4))
+#define OFFSET_FROM_BIT(a) (a%(8*4))
+
 typedef struct
 {
 	uint32_t present	: 1;
-	uint32_t rw		: 1;
+	uint32_t rw			: 1;
 	uint32_t user		: 1;
 	uint32_t accessed	: 1;
 	uint32_t dirty		: 1;
