+++ src/kernel/shell.c
@@ -1,3 +1,8 @@
+// Basic shell and commands kernel module
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "system.h"
 #include "../libc/stdio.h"
 #include "../libc/string.h"
@@ -6,8 +11,7 @@
 
 #define BUFFER_SIZE 256
 #define MAX_COMMANDS 16
-#define MAX_ARGS 64
-
+#define MAX_ARGS     64
 
 char* ascii_title =
 "\n"
@@ -16,7 +20,6 @@ char* ascii_title =
 "Author: @xamidev - star the repo for a cookie!\n"
 "----------------------------------------------\n"
 "\n";
-
 
 typedef void (*command_func_t)(int argc, char *argv[]);
