+++ src/kernel/sysinfo.c
@@ -1,3 +1,8 @@
+// System information kernel module
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "../libc/stdio.h"
 #include "../libc/string.h"
