+++ src/kernel/sysinfo.h
@@ -1,3 +1,8 @@
+// System information kernel module header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef SYSINFO_H
 #define SYSINFO_H
