+++ src/kernel/system.c
@@ -1,3 +1,8 @@
+// System utilities and routines kernel module
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "system.h"
 #include <stdint.h>
