+++ src/kernel/system.h
@@ -1,3 +1,8 @@
+// System utilities and routines kernel module header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef SYSTEM_H
 #define SYSTEM_H
