+++ src/libc/crypto.c
@@ -1,3 +1,8 @@
+// Cryptography routines for blankos/libc
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "crypto.h"
 #include <stdint.h>
