+++ src/libc/crypto.h
@@ -1,3 +1,8 @@
+// Cryptography routines for blankos/libc header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef CRYPTO_H
 #define CRYPTO_H
