+++ src/libc/ctype.c
@@ -1,3 +1,7 @@
+// Ctype implementation for blankos/libc
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
 
 #include <stdint.h>
 #include <stdbool.h>
@@ -6,7 +10,6 @@ bool isdigit(char c)
 {
 	return c >= '0' && c <= '9';
 }
-
 
 bool isspace(char c)
 {
