+++ src/libc/ctype.h
@@ -1,3 +1,8 @@
+// Ctype implementation for blankos/libc header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef CTYPE_H
 #define CTYPE_H
