+++ src/libc/stdio.h
@@ -1,17 +1,24 @@
-#ifndef INCLUDE_STDIO_H
-#define INCLUDE_STDIO_H
+// Standard input/output implementation for blankos/libc header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
+#ifndef STDIO_H
+#define STDIO_H
 
 #include <stdint.h>
 #include <stdbool.h>
 
-#define FB_GREEN      2
-#define FB_DARK_GREY  8
+#define FB_GREEN      	  2
+#define FB_DARK_GREY  	  8
 
 #define FB_CMD_PORT       0x3D4
 #define FB_DATA_PORT      0x3D5
 #define FB_HIGH_BYTE_CMD  14
 #define FB_LOW_BYTE_CMD   15
 
+#define CURSOR_WIDTH  	  8
+#define CURSOR_HEIGHT     16
 
 void draw_cursor(uint32_t color);
 void erase_cursor();
@@ -27,24 +34,23 @@ unsigned int getcolor(int x, int y);
 void putc(char c);
 void colorputc(char c, uint32_t fg, uint32_t bg);
 
-#define PRINTF_STATE_START 		0
+#define PRINTF_STATE_START 			0
 #define PRINTF_STATE_LENGTH 		1
-#define PRINTF_STATE_SHORT 		2
-#define PRINTF_STATE_LONG 		3
-#define PRINTF_STATE_SPEC 		4
-#define PRINTF_STATE_WIDTH 		5
+#define PRINTF_STATE_SHORT 			2
+#define PRINTF_STATE_LONG 			3
+#define PRINTF_STATE_SPEC 			4
+#define PRINTF_STATE_WIDTH 			5
 
 #define PRINTF_LENGTH_START 		0
 #define PRINTF_LENGTH_SHORT_SHORT 	1
 #define PRINTF_LENGTH_SHORT 		2
-#define PRINTF_LENGTH_LONG 		3
+#define PRINTF_LENGTH_LONG 			3
 #define PRINTF_LENGTH_LONG_LONG 	4
 
 void printf(const char* fmt, ...);
 int* printf_number(int* argp, int length, bool sign, int radix, int width, char pad_char);
 int getch();
 void get_input(char *buffer, int size);
-
 void dtostrf(double val, char *buffer, int precision);
 
 enum Colors
