+++ src/libc/string.c
@@ -1,3 +1,8 @@
+// String operations implementation for blankos/libc
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include <stdint.h>
 #include "../kernel/system.h"
