+++ src/programs/ciphers.h
@@ -1,3 +1,8 @@
+// Cipher programs header
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #ifndef CIPHERS_H
 #define CIPHERS_H
