+++ src/programs/conway.c
@@ -1,3 +1,8 @@
+// Conway's Game of Life program
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include "conway.h"
 #include "../libc/stdio.h"
 #include "../kernel/system.h"
