+++ src/programs/math.c
@@ -1,4 +1,7 @@
-// Math expression lexer and parser
+// Basic math expression lexer and parser program
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
 
 #include <stdint.h>
 #include "../kernel/system.h"
