+++ src/programs/primes.c
@@ -1,3 +1,8 @@
+// Prime number computation program
+// Author: xamidev
+// Licensed under the Unlicense. See the repo below.
+// https//github.com/xamidev/blankos
+
 #include <stdint.h>
 #include "../libc/stdio.h"
 #include "../kernel/system.h"
